# Optimize xml comments