Optimized Dummy website allocations from 4.6GB to 245Mb and improved runtime performance from 4.6seconds to 637ms.

> 
| Method                | Mean     | Error   | StdDev  | Gen0       | Gen1      | Allocated |
|---------------------- |---------:|--------:|--------:|-----------:|----------:|----------:|
| DotSwashbuckleOpenApi | 637.7 ms | 5.32 ms | 4.16 ms | 15000.0000 | 4000.0000 | 245.53 MB |